app/requirements/dev.txt
-r common.txt
black
ruff
mypy
tox
django-debug-toolbar
pytest-django
pytest-randomly
pytest-cov
factory-boy
pytest-structlog
sphinx
